Use explicit localhost in htex_local_alternate tests
I was seeing delays in some circumstances of up to 30 seconds when running
tests with this configuration.

I suspected that this was due to automatic address configuration so I've
removed that and now use an explicit `localhost`.

This moves test run time to a consistent 11s against master on my laptop.

I'm not sure if this should be merged to master, as it stops the auto
address configuration from being tested.