site tests were running in --config local mode
i dont think this is appropriate with automatic site detection, if site detection is intended to fail (as I want in a subsequent PR) rather than default to submit host htex.

so this patch should make those site tests/ tests only run in site test mode

This patch really needs to clarify *when* site_tests tests are intended to run, to make things *more* coherent rather than less coherent,
especially in the context of the subsequent PR which raises an exception.

Another question to ask, for example, perhaps deeper, is how do I run these site_tests with a config of my choice rather than with the config that the system has chosen for me?
Or can I not do that? (I can't at the moment...)
One option is that they *should* be able to get a fresh config from the --config option every time.
Another option is to accept that these tests are magical and you cannot specify the config you want to test.