Add some pytests around irregular shutdown patterns
Ideas:

ctrl-C style: kill one of the monitoring processes and check this doesn't hang. This hangs right now, I think, and I would like it to not hang.

ctrl-C style: kill htex interchange before DFK shutdown

sigterm: ctrl-c uses sigint. repeat the same tests, but with sigterm, which is handled differently by python by default (maybe no atexit handling?)

delayed dependency: long task - next shut down the DFK - next that long task completes (perhaps by a flag) and that causes a dependency to attempt to start, even though the DFK is shut down. What should the behaviour even be for this?


when pressing ctrl-c in pytest, prior to this PR, it would hang as wait_for_task_completion would be called in the unwinding of the fixture stack. However, because
ctrl-c was pressed, tasks wouldn't be expected to all complete.

This PR changes that to assert that all tasks are completed, rather than *waiting* for all tasks to complete.

A non-finished task now gives a test error - which is arguably better anyway because it more aggressively flushes out tests that do not perform a complete shutdown.
This means that pressing ctrl-C in a pytest leads to an assertion error; when previously it led to a hang.
This is related to however pytest does its ctrl-c handling.