Switch to module-based singleton for ParslSerializer
Deserialization of a singleton is conceptually awkward.
Prior to this patch, the state of the singleton is
overwritten by the state supplied in the serialized
singleton, which can produce surprising cache behaviour:
cached functions would "disappear" from the cache as
the registered DillSerializer is overwritten with a different
newly deserializer DillSerializer.

TODO: investigate what happens when this module is serialized and deserialized now - what does deserializing a module look like at the remote end?