Move monitoring wrapper code into its own module
This is to avoid capturing so much in the closure of the monitoring wrapper,
which prior to this patch, captured a lot of surprising stuff including a
serialization of the shouldn't-be-serialised ParslSerializer state.

Because this (hopefully) [TODO: write numbers] gives a smaller closure, the
performance of task launching might be slightly faster?

on my laptop, 166 launches/sec before, 230 launches/sec after