hack to stop import of `parsl` importing so much, so that when it is imported in a WQ worker implicitly by importing a parsl.* submodule, all this symbol re-export doesn't happen.
this isn't intended as a serious feature right now - more for benchmarking if it changes behaviour

could play with some kind of lazy importing to make it happen automatically too: override the code that looks up names and submodules, and perform the import/set at that point, instead of at the point that __init__.py is loaded/imported.

for lazy loading, see this:
https://scientific-python.org/specs/spec-0001/

for wanting to do this for very similar motivations

on my laptop, this shifts "import parsl" time with no further activity from 0.4s to 0.04 seconds.